requirements.txt
libcst
fake-bpy-module-4.1
black
isort
ruff
pytest
blender-stubs
